function 1 in progress

+++ lab4/mouse.h
@@ -7,7 +7,7 @@ extern struct packet mouse_packet;
 uint32_t mouse_status;
 extern int bytes_count;
 int bit_no_global_mouse;
-uint32_t packet_byte;
+uint8_t packet_byte;
 
 int(mouse_subscribe_int)(uint8_t *bit_no);
 int(mouse_unsubscribe_int)();
